Move kms client to dedicated module
